refactor(react): remove useStream hook
Move the useStream hook into the connect module, and do not expose it through any index file.

BREAKING CHANGE: Removes useStream hook. Use connect instead.
