+++ src/react/index.ts
@@ -1,2 +1,1 @@
-export { useStream, NOT_YET_EMITTED } from './useStream';
 export { connect } from './connect';
